+++ src/const.js
@@ -1,0 +1,47 @@
+const CITIES = [
+  `Amsterdam`,
+  `Geneva`,
+  `Chamonix`,
+  `Saint Petersburg`,
+  `Paris`,
+  `New York`
+];
+
+const SENTENCE = [
+  `Lorem ipsum dolor sit amet, consectetur adipiscing elit.`,
+  `Cras aliquet varius magna, non porta ligula feugiat eget.`,
+  `Fusce tristique felis at fermentum pharetra.`,
+  `Aliquam id orci ut lectus varius viverra.`,
+  `Nullam nunc ex, convallis sed finibus eget, sollicitudin eget ante.`,
+  `Phasellus eros mauris, condimentum sed nibh vitae, sodales efficitur ipsum.`,
+  `Sed blandit, eros vel aliquam faucibus, purus ex euismod diam, eu luctus nunc ante ut dui.`,
+  `Sed sed nisi sed augue convallis suscipit in sed felis.`,
+  `Aliquam erat volutpat.`,
+  `Nunc fermentum tortor ac porta dapibus.`,
+  `In rutrum ac purus sit amet tempus.`
+];
+
+const OFFERS_MAP = {
+  luggage: {
+    name: `Add luggage`,
+    price: 30
+  },
+  comfort: {
+    name: `Switch to comfort`,
+    price: 100
+  },
+  meal: {
+    name: `Add meal`,
+    price: 15
+  },
+  seats: {
+    name: `Choose seats`,
+    price: 5
+  },
+  train: {
+    name: `Travel by train`,
+    price: 40
+  }
+};
+
+export {CITIES, SENTENCE, OFFERS_MAP};
